requirements.txt
apache-airflow-providers-google ==10.15.0
apache-airflow-providers-snowflake ==5.3.0
snowflake-connector-python ==3.7.0
snowflake-sqlalchemy ==1.5.1
google-cloud-storage==2.14.0
pandas==2.0.3 

